models/requirements.txt
torch==2.8.0
transformers==4.57.1
diffusers==0.35.2
accelerate==1.11.0
xformers==0.0.32.post2
huggingface-hub==0.36.0
numpy==2.3.4
Pillow==12.0.0
tqdm==4.67.1

